pyproject.toml
[project]
name = "directmultistep"
version = "0.9.0"
dependencies = [
    "numpy==1.26.4",
    "pandas==1.5.3",
    "pdoc3==0.11.5",
    "plotly==5.24.1",
    "lightning==2.2.5",
    "pyyaml==6.0.1",
    "rdkit==2023.9.3",
    "torch==2.3.0",
    "torchmetrics==1.6.0",
    "tqdm==4.67.1",
    "svgwrite==1.4.3",
    "svglib==1.5.1",
    "tomli>=2.2.1",
]

[tool.setuptools]
packages = ["directmultistep"]
package-dir = {"" = "src"}



[project.optional-dependencies]
dev = [
    "ipykernel>=6.29.5",
    "nbformat>=5.10.4",
    "rich>=13.9.4",
    "kaleido==0.2.1",
    "pre-commit==4.0.1",
    "mkdocs==1.6.1",
    "mkdocstrings-python==1.12.2",
    "mkdocs-material==9.5.49",
    "pytest==8.3.4",
    "ruff==0.4.7",
    "mypy==1.13.0",
    "isort==5.13.2",
    "typing-extensions==4.12.2",
    "mypy-extensions==1.0.0",
    "types-pyyaml",
    "types-tqdm",
    "types-requests"
]

[tool.mypy]
strict = true
ignore_missing_imports = true
exclude = ["tests"]
disable_error_code = ["unused-ignore"]

[[tool.mypy.overrides]]
module = ["rdkit-stubs.*", "rdkit.*"]
ignore_errors = true

[tool.ruff]
line-length = 120

[tool.logging]
version = 1
disable_existing_loggers = false

[tool.logging.formatters.standard]
format = "%(asctime)s [%(levelname)s] %(name)s: %(message)s"
datefmt = "%Y-%m-%d %H:%M:%S"

[tool.logging.handlers.console]
class = "logging.StreamHandler"
formatter = "standard"
stream = "ext://sys.stdout"

[tool.logging.loggers.directmultistep]
handlers = ["console"]
propagate = false
level = "INFO"
